chore: rename pkg/hash to pkg/pwhash
This avoids the conflict with the stdlib hash package.
